% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-edit.R
\name{schema_add_def}
\alias{schema_add_def}
\title{Add a schema definition}
\usage{
schema_add_def(x, name, value, overwrite = FALSE)
}
\arguments{
\item{x}{A \code{SchemaDoc}.}

\item{name}{Definition name to add.}

\item{value}{Schema fragment using the same list syntax accepted by
\code{schema_doc()}, or a fragment produced by helpers such as \code{schema_check()},
to store in \verb{$defs}.}

\item{overwrite}{Logical flag indicating whether an existing definition of the
same name should be replaced.}
}
\value{
A modified \code{SchemaDoc}.
}
\description{
Add a schema definition
}
\examples{
schema <- schema_doc(schema_check("string"))
schema <- schema_add_def(schema, "text", schema_check("string"))
schema

names(as.list(schema)$`$defs`)

}
