% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemate-package.R
\docType{package}
\name{schemate-package}
\alias{schemate}
\alias{schemate-package}
\title{schemate: Checkmate-first schema inference, editing, JSON IO, and validation}
\description{
\code{schemate} provides a compact schema DSL for R data. It can infer schemas
from example objects, edit schema documents, read and write schema JSON, and
validate inputs against compiled schemas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hongyuanjia.github.io/schemate/}
  \item Report bugs at \url{https://github.com/hongyuanjia/schemate/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID}) [copyright holder]

}
\keyword{internal}
