% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-edit.R
\name{schema_del_def}
\alias{schema_del_def}
\title{Delete a schema definition}
\usage{
schema_del_def(x, name, error_if_missing = TRUE)
}
\arguments{
\item{x}{A \code{SchemaDoc}.}

\item{name}{Definition name to remove.}

\item{error_if_missing}{Logical flag indicating whether a missing definition
should raise an error.}
}
\value{
A modified \code{SchemaDoc}.
}
\description{
Delete a schema definition
}
