% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-edit.R
\name{schema_del_position}
\alias{schema_del_position}
\title{Delete a position schema from an unnamed container node}
\usage{
schema_del_position(x, index, path = "$", error_if_missing = TRUE)
}
\arguments{
\item{x}{A \code{SchemaDoc}.}

\item{index}{1-based position index to remove.}

\item{path}{Path to the target unnamed container node. Use \code{$} for the root
node.}

\item{error_if_missing}{Logical flag indicating whether a missing position
schema should raise an error.}
}
\value{
A modified \code{SchemaDoc}.
}
\description{
Delete a position schema from an unnamed container node
}
